+++ front_end/src/components/AsyncContainer/AsyncContainer.vue
