+++ front_end/src/components/card/StellaCard/StellaCard.vue
