+++ front_end/src/components/common/GridCardListContainer/GridCardListContainer.vue
