+++ front_end/src/components/skeleton/CardSkeleton/CardSkeleton.vue
