+++ front_end/src/components/views/home/IntroductionSection.vue
@@ -1,5 +1,5 @@
 <script setup lang="ts">
-import StellaCard from '@/components/StellaCard/StellaCard.vue';
+import StellaCard from '@/components/card/StellaCard/StellaCard.vue';
 import useInView from '@/composables/useInView';
 import { useStaggeredAnimation } from '@/composables/useStaggeredAnimation';
 import { ref } from 'vue';
